+++ src/main/java/com/project/yogerOrder/payment/controller/PaymentController.java
@@ -1,7 +1,5 @@
 package com.project.yogerOrder.payment.controller;
 
-import com.project.yogerOrder.order.config.OrderTopic;
-import com.project.yogerOrder.order.event.OrderCanceledEvent;
 import com.project.yogerOrder.payment.dto.request.PartialRefundRequestDTOs;
 import com.project.yogerOrder.payment.dto.request.PortOnePaymentWebhookRequestDTO;
 import com.project.yogerOrder.payment.dto.request.VerifyPaymentRequestDTO;
@@ -10,8 +8,6 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.kafka.annotation.KafkaListener;
-import org.springframework.kafka.support.Acknowledgment;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -36,13 +32,5 @@ public class PaymentController {
         paymentService.productsExpiration(partialRefundRequestDTOs);
 
         return new ResponseEntity<>(HttpStatus.OK);
-    }
-
-    @KafkaListener(topics = OrderTopic.CANCELED, groupId = "payment-group",
-            containerFactory = "orderCanceledFactory")
-    public void orderCanceled(OrderCanceledEvent event, Acknowledgment acknowledgment) {
-        paymentService.orderCanceled(event.orderId());
-
-        acknowledgment.acknowledge();
     }
 }
