+++ src/main/java/com/project/yogerOrder/payment/event/consumer/PaymentEventConsumer.java
@@ -1,0 +1,24 @@
+package com.project.yogerOrder.payment.event.consumer;
+
+import com.project.yogerOrder.order.config.OrderTopic;
+import com.project.yogerOrder.order.event.OrderCanceledEvent;
+import com.project.yogerOrder.payment.service.PaymentService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.kafka.annotation.KafkaListener;
+import org.springframework.kafka.support.Acknowledgment;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class PaymentEventConsumer {
+
+    private final PaymentService paymentService;
+
+    @KafkaListener(topics = OrderTopic.CANCELED, groupId = "payment-group",
+            containerFactory = "orderCanceledFactory")
+    public void orderCanceled(OrderCanceledEvent event, Acknowledgment acknowledgment) {
+        paymentService.orderCanceled(event.orderId());
+
+        acknowledgment.acknowledge();
+    }
+}
